requirements.txt
alpaca-py==0.40.1
annotated-types==0.7.0
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
msgpack==1.1.1
numpy==2.3.0
pandas==2.3.0
pyarrow==20.0.0
pydantic==2.11.7
pydantic-core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.4
six==1.17.0
sseclient-py==1.8.0
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
boto3
s3fs
